refactor(approval-gate): extract resolve.rs from lib.rs
Fifth step of breaking lib.rs into focused modules. This commit
lifts the approval::resolve flow into its own module.

Moved (lib.rs -> resolve.rs):
- handle_lookup_record()
- handle_resolve()
- cascade_allow_for_session() (the always=true sweep)
- approve_and_execute() (shared by user-driven allow + cascade)

handle_resolve and handle_lookup_record stay public via the crate
root re-exports. approve_and_execute and cascade_allow_for_session
stay private to the resolve module.

Pure move, no behavior change. 141 approval-gate tests pass.
